requirements.txt
torch
torchvision
pillow
matplotlib
scikit-learn
pandas
numpy
nonechucks

